ai-models/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.9.2
pydantic-settings==2.11.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.5.2
sentence-transformers==3.1.1
nltk==3.9.1
python-dotenv==1.0.1
requests==2.32.3
google-generativeai==0.8.3
